app/backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
pydantic>=2.6.0
pydantic-settings>=2.2.0
mlflow>=2.17.0
lightgbm>=4.3.0
numpy>=1.26.0
pandas>=2.2.0
boto3>=1.34.0
redis>=5.0.0
python-dotenv>=1.0.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
bcrypt>=4.0.1,<4.1
websockets>=12.0
prometheus-fastapi-instrumentator>=7.0.0
sqlalchemy[asyncio]>=2.0.0
asyncpg>=0.29.0


